feat: Add Logbook configuration for HTTP request/response logging and update AccountTypeRepository ID type

+++ src/main/java/com/matjazt/netmon2/repository/AccountTypeRepository.java
@@ -14,7 +14,7 @@ import java.util.Optional;
  * change often.
  */
 @Repository
-public interface AccountTypeRepository extends JpaRepository<AccountTypeEntity, Long> {
+public interface AccountTypeRepository extends JpaRepository<AccountTypeEntity, Integer> {
 
     /**
      * Find account type by name (e.g., "Admin", "Viewer", "MonitoringDevice")
